+++ src/index.ts
@@ -1,3 +1,4 @@
 import { stemFrenchMinimal } from "./stemFrenchMinimal";
+import { stemFrenchLight } from "./stemFrenchLight";
 
-export { stemFrenchMinimal };
+export { stemFrenchMinimal, stemFrenchLight };